Complete re-architecture of bgee to allow testing
